Respond to Tristan comments
