+++ test_runner/fixtures/neon_fixtures.py
@@ -87,7 +87,7 @@ from fixtures.utils import (
 )
 from fixtures.utils import AuxFileStore as AuxFileStore  # reexport
 
-from .neon_api import NeonAPI, connection_parameters_to_env
+from .neon_api import NeonAPI, NeonApiEndpoint
 
 """
 This file contains pytest fixtures. A fixture is a test resource that can be
@@ -3158,65 +3158,16 @@ class RemotePostgres(PgProtocol):
         pass
 
 
-class RemoteNeonApiEndpoint:
-    def __init__(self, neon_api: NeonAPI, pg_version: PgVersion, project_id: Optional[str]):
-        self.neon_api = neon_api
-        if project_id is None:
-            project = neon_api.create_project(pg_version)
-            neon_api.wait_for_operation_to_finish(project["project"]["id"])
-            self.project_id = project["project"]["id"]
-            self.endpoint_id = project["endpoints"][0]["id"]
-            self.connstr = project["connection_uris"][0]["connection_uri"]
-            self.pgbench_env = connection_parameters_to_env(
-                project["connection_uris"][0]["connection_parameters"]
-            )
-            self.is_new = True
-        else:
-            project = neon_api.get_project_details(project_id)
-            if int(project["project"]["pg_version"]) != int(pg_version):
-                raise Exception(
-                    f"A project with the provided ID exists, but it's not of the specified version (expected {pg_version}, got {project['project']['pg_version']})"
-                )
-            self.project_id = project_id
-            eps = neon_api.get_endpoints(project_id)["endpoints"]
-            self.endpoint_id = eps[0]["id"]
-            self.connstr = neon_api.get_connection_uri(project_id, endpoint_id=self.endpoint_id)[
-                "uri"
-            ]
-            pw = self.connstr.split("@")[0].split(":")[-1]
-            self.pgbench_env = {
-                "PGHOST": eps[0]["host"],
-                "PGDATABASE": "neondb",
-                "PGUSER": "neondb_owner",
-                "PGPASSWORD": pw,
-            }
-            self.is_new = False
-
-    def restart(self):
-        self.neon_api.restart_endpoint(self.project_id, self.endpoint_id)
-        self.neon_api.wait_for_operation_to_finish(self.project_id)
-
-    def get_synthetic_storage_size(self) -> int:
-        return int(
-            self.neon_api.get_project_details(self.project_id)["project"]["synthetic_storage_size"]
-        )
-
-
 @pytest.fixture(scope="function")
-def benchmark_project_pub(neon_api: NeonAPI, pg_version: PgVersion) -> RemoteNeonApiEndpoint:
+def benchmark_project_pub(neon_api: NeonAPI, pg_version: PgVersion) -> NeonApiEndpoint:
     project_id = os.getenv("BENCHMARK_PROJECT_ID_PUB")
-    return RemoteNeonApiEndpoint(neon_api, pg_version, project_id)
+    return NeonApiEndpoint(neon_api, pg_version, project_id)
 
 
 @pytest.fixture(scope="function")
-def benchmark_project_sub(neon_api: NeonAPI, pg_version: PgVersion) -> RemoteNeonApiEndpoint:
+def benchmark_project_sub(neon_api: NeonAPI, pg_version: PgVersion) -> NeonApiEndpoint:
     project_id = os.getenv("BENCHMARK_PROJECT_ID_SUB")
-    return RemoteNeonApiEndpoint(neon_api, pg_version, project_id)
-
-
-@pytest.fixture(scope="function")
-def benchmark_connstr_sub() -> Optional[str]:
-    return os.getenv("BENCHMARK_CONNSTR_SUB")
+    return NeonApiEndpoint(neon_api, pg_version, project_id)
 
 
 @pytest.fixture(scope="function")
